+++ CPAC/utils/configuration.py
@@ -82,10 +82,10 @@ class Configuration:
     'new_pipeline2'
     """
     def __init__(self, config_map=None):
+        from click import BadParameter
         from CPAC.pipeline.schema import schema
         from CPAC.utils.utils import load_preconfig, lookup_nested_value, \
             update_nested_dict
-        from optparse import OptionError
 
         if config_map is None:
             config_map = {}
@@ -96,8 +96,8 @@ class Configuration:
         if base_config not in ['default', 'default_pipeline']:
             try:
                 base_config = load_preconfig(base_config)
-            except OptionError:
-                base_config = base_config
+            except BadParameter:
+                pass
             from_config = yaml.safe_load(open(base_config, 'r'))
             config_map = update_nested_dict(
                 Configuration(from_config).dict(), config_map)